Fix: 공동구매 서비스 테스트 수정
related to #14

+++ src/main/java/javachip/entity/GroupBuy.java
@@ -21,7 +21,7 @@ public class GroupBuy {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "groupbuy_seq_gen")
     @SequenceGenerator(name = "groupbuy_seq_gen", sequenceName = "GROUPBUY_SEQ", allocationSize = 1)
     @Column(name = "ID")
-    private Long id; // groupBuyId
+    private Integer id; // groupBuyId
 
     @OneToOne
     @JoinColumn(name = "PRODUCT_ID", nullable = false, unique = true)
